Split more code
