+++ Python/Product/Analysis/LanguageServer/Handlers/WorkspaceSymbolsHandler.cs
